+++ ufotest/scripts/pcie_init.sh
@@ -6,7 +6,7 @@ if [ -z "$device" ]; then
   echo 1 > /sys/bus/pci/rescan
   exit
 else
-  echo "Xilin is located at: " $device
+  echo "Xilinx is located at: " $device
 fi
 echo "remove devices"
 # echo 1 > /sys/bus/pci/devices/0000\:01\:00.0/remove
@@ -17,7 +17,7 @@ echo "rescan"
 echo 1 > /sys/bus/pci/rescan
 sleep 1
 echo "remove driver"
-rmmod pciDriver 
+rmmod pciDriver
 sleep 1
 echo "instantiate driver"
 modprobe pciDriver
@@ -30,12 +30,12 @@ pci -i
 sleep .1
 pci -r 9000
 echo "set bus master dma"
-# dev=04:00.0  
-dev=$device  
+# dev=04:00.0
+dev=$device
 echo Enabling bus mastering on device $dev
 setpci -s $dev 4.w=0x07
 
 export LD_LIBRARY_PATH=/usr/lib
 export PCILIB_MODEL=ipedma
 
-./dma.sh
+dma.sh